+++ app/public/js/querySearchTypeahead.js
@@ -3,6 +3,7 @@ const NHSUK = NHSUK || {};
 
 NHSUK.queryTypeahead = ((global) => {
   const $ = global.jQuery;
+  const mainId = '#main-content';
   const maxResultCount = 10;
   const searchField = '#query';
   // TODO: Ideally these values will come from environment variables
@@ -11,7 +12,7 @@ NHSUK.queryTypeahead = ((global) => {
   const indexName = 'organisationlookup3-index';
   const suggestHost = 'nhsukpoc.search.windows.net';
   const suggesterName = 'orgname-suggester';
-  const searchUrl = './results?type=iapt';
+  const resultsUrl = './results?type=iapt';
   const suggestUrl = `https://${suggestHost}/indexes/${indexName}/docs/suggest?api-version=${apiVersion}`;
 
   const suggestions = new Bloodhound({
@@ -43,6 +44,12 @@ NHSUK.queryTypeahead = ((global) => {
     },
   });
 
+  function generateIAPTResultsUrl(data) {
+    const ccgid = encodeURIComponent(data.CCG[0]);
+    const gpname = encodeURIComponent(data.OrganisationName);
+    return `${resultsUrl}&ccgid=${ccgid}&gpquery=${gpname}&gpname=${gpname}&origin=search`;
+  }
+
   function init() {
     suggestions.initialize();
 
@@ -68,35 +75,33 @@ NHSUK.queryTypeahead = ((global) => {
         header: '<li class="c-search-menu__prepend">Search suggestions</li>',
         notFound: '<li class="c-search-menu__nosuggestions">No suggestions</li>',
         suggestion: (data) => {
-          const query = encodeURIComponent(data.CCG[0]);
-          const gpname = encodeURIComponent(data.OrganisationName);
-          return `<li><a href="${searchUrl}&query=${query}&gpquery=${gpname}&gpname=${gpname}&origin=search">${data.OrganisationName}, ${data.City}, ${data.Postcode}</a></li>`;
+          const link = generateIAPTResultsUrl(data);
+          return `<li><a href="${link}">${data.OrganisationName}, ${data.City}, ${data.Postcode}</a></li>`;
         },
       },
     })
       .bind('typeahead:open', () => {
-        const val = $(searchField).typeahead('val');
-        const value = $(searchField).attr('value');
+        const $searchField = $(searchField);
+        const val = $searchField.typeahead('val');
+        const value = $searchField.attr('value');
 
         if (val === value) {
-          $(searchField).typeahead('val', value);
-        }
-        if (val.toLowerCase() === 'enter a search term') {
-          $(searchField).typeahead('val', '');
+          $searchField.typeahead('val', value);
         }
       })
       .bind('typeahead:render', () => {
         const $searchField = $(searchField);
-        $('.c-search-menu__results').wrapInner('<ul class="c-search-menu__list"></ul>');
-        $('.c-search-menu__list').css('width', $searchField.outerWidth());
-        $('.c-search-menu').insertAfter($searchField);
+        $(`${mainId} .c-search-menu__results`).wrapInner('<ul class="c-search-menu__list"></ul>');
+        $(`${mainId} .c-search-menu__list`).css('width', $searchField.outerWidth());
+        $(`${mainId} .c-search-menu`).insertAfter($searchField);
       })
-      .bind('typeahead:close', () => {
-        $('.c-search__input').removeClass('c-search__input--dropdown');
-        $('.c-search__submit').removeClass('c-search__submit--dropdown');
-      })
-      .bind('typeahead:idle', () => {
-        $('.c-search-menu__list').hide();
+      .bind('typeahead:select', (e, o) => {
+        $('#type').val('iapt');
+        $('#ccgid').val(o.CCG[0]);
+        $('#gpname').val(o.OrganisationName);
+        $('#gpquery').val(o.OrganisationName);
+        $('#origin').val('search');
+        $('main form').submit();
       });
   }
 
@@ -107,7 +112,8 @@ NHSUK.queryTypeahead = ((global) => {
 
 $(() => {
   NHSUK.queryTypeahead.init();
+  const mainId = '#main-content';
   // hide the extra input field created by typeahead to screen readers
-  $('.c-search__input--shadow').attr('aria-hidden', 'true').addClass('visually-hidden');
-  $('.c-search__input.tt-input').attr('role', 'textbox');
+  $(`${mainId} .c-search__input--shadow`).attr('aria-hidden', 'true').addClass('visually-hidden');
+  $(`${mainId} .c-search__input.tt-input`).attr('role', 'textbox');
 });